+++ Source/Scene/Cesium3DTile.js
@@ -23,6 +23,7 @@ import RequestState from "../Core/RequestState.js";
 import RequestType from "../Core/RequestType.js";
 import Resource from "../Core/Resource.js";
 import RuntimeError from "../Core/RuntimeError.js";
+import Cesium3DContentGroup from "./Cesium3DContentGroup.js";
 import Cesium3DTileContentFactory from "./Cesium3DTileContentFactory.js";
 import Cesium3DTileContentState from "./Cesium3DTileContentState.js";
 import Cesium3DTileContentType from "./Cesium3DTileContentType.js";
@@ -1317,7 +1318,13 @@ function makeContent(tile, arrayBuffer) {
     content.metadata = findContentMetadata(tileset, contentHeader);
   }
 
-  content.groupMetadata = findGroupMetadata(tileset, contentHeader);
+  const groupMetadata = findGroupMetadata(tileset, contentHeader);
+  if (defined(groupMetadata)) {
+    content.group = new Cesium3DContentGroup({
+      metadata: groupMetadata,
+    });
+  }
+
   return content;
 }