+++ Source/Scene/Cesium3DTileContent.js
@@ -245,7 +245,7 @@ Object.defineProperties(Cesium3DTileContent.prototype, {
   },
 
   /**
-   * Gets the group metadata for this content if the content has metadata (3D Tiles 1.1) or
+   * Gets the group for this content if the content has metadata (3D Tiles 1.1) or
    * if it uses the <code>3DTILES_metadata</code> extension. If neither are present,
    * this property should be undefined.
    * <p>
@@ -253,12 +253,12 @@ Object.defineProperties(Cesium3DTileContent.prototype, {
    * not part of the public Cesium API.
    * </p>
    *
-   * @type {GroupMetadata|undefined}
+   * @type {Cesium3DTileContentGroup|undefined}
    *
    * @private
    * @experimental This feature is using part of the 3D Tiles spec that is not final and is subject to change without Cesium's standard deprecation policy.
    */
-  groupMetadata: {
+  group: {
     // eslint-disable-next-line getter-return
     get: function () {
       DeveloperError.throwInstantiationError();