+++ Source/Scene/ModelExperimental/ModelExperimental3DTileContent.js
@@ -1,5 +1,6 @@
 import Axis from "../Axis.js";
 import Color from "../../Core/Color.js";
+import combine from "../../Core/combine.js";
 import defined from "../../Core/defined.js";
 import destroyObject from "../../Core/destroyObject.js";
 import Pass from "../../Renderer/Pass.js";
@@ -27,7 +28,7 @@ export default function ModelExperimental3DTileContent(
 
   this._model = undefined;
   this._metadata = undefined;
-  this._groupMetadata = undefined;
+  this._group = undefined;
 }
 
 Object.defineProperties(ModelExperimental3DTileContent.prototype, {
@@ -128,12 +129,12 @@ Object.defineProperties(ModelExperimental3DTileContent.prototype, {
     },
   },
 
-  groupMetadata: {
+  group: {
     get: function () {
-      return this._groupMetadata;
+      return this._group;
     },
     set: function (value) {
-      this._groupMetadata = value;
+      this._group = value;
     },
   },
 });
@@ -194,6 +195,7 @@ ModelExperimental3DTileContent.prototype.update = function (
   model.pointCloudShading = tileset.pointCloudShading;
   model.featureIdLabel = tileset.featureIdLabel;
   model.instanceFeatureIdLabel = tileset.instanceFeatureIdLabel;
+  model.lightColor = tileset.lightColor;
   model.backFaceCulling = tileset.backFaceCulling;
   model.shadows = tileset.shadows;
   model.showCreditsOnScreen = tileset.showCreditsOnScreen;
@@ -218,23 +220,17 @@ ModelExperimental3DTileContent.fromGltf = function (
 ) {
   const content = new ModelExperimental3DTileContent(tileset, tile, resource);
 
-  const modelOptions = {
+  const additionalOptions = {
     gltf: gltf,
-    cull: false, // The model is already culled by 3D Tiles
-    releaseGltfJson: true, // Models are unique and will not benefit from caching so save memory
-    opaquePass: Pass.CESIUM_3D_TILE, // Draw opaque portions of the model during the 3D Tiles pass
     basePath: resource,
-    modelMatrix: tile.computedTransform,
-    upAxis: tileset._gltfUpAxis,
-    forwardAxis: Axis.X,
-    incrementallyLoadTextures: false,
-    customShader: tileset.customShader,
-    content: content,
-    colorBlendMode: tileset.colorBlendMode,
-    colorBlendAmount: tileset.colorBlendAmount,
-    backFaceCulling: tileset.backFaceCulling,
-    shadows: tileset.shadows,
   };
+
+  const modelOptions = makeModelOptions(
+    tileset,
+    tile,
+    content,
+    additionalOptions
+  );
   content._model = ModelExperimental.fromGltf(modelOptions);
   return content;
 };
@@ -248,22 +244,18 @@ ModelExperimental3DTileContent.fromB3dm = function (
 ) {
   const content = new ModelExperimental3DTileContent(tileset, tile, resource);
 
-  const modelOptions = {
+  const additionalOptions = {
     arrayBuffer: arrayBuffer,
     byteOffset: byteOffset,
     resource: resource,
-    cull: false, // The model is already culled by 3D Tiles
-    releaseGltfJson: true, // Models are unique and will not benefit from caching so save memory
-    opaquePass: Pass.CESIUM_3D_TILE, // Draw opaque portions of the model during the 3D Tiles pass
-    modelMatrix: tile.computedTransform,
-    upAxis: tileset._gltfUpAxis,
-    forwardAxis: Axis.X,
-    incrementallyLoadTextures: false,
-    customShader: tileset.customShader,
-    content: content,
-    colorBlendMode: tileset.colorBlendMode,
-    colorBlendAmount: tileset.colorBlendAmount,
   };
+
+  const modelOptions = makeModelOptions(
+    tileset,
+    tile,
+    content,
+    additionalOptions
+  );
   content._model = ModelExperimental.fromB3dm(modelOptions);
   return content;
 };
@@ -277,22 +269,18 @@ ModelExperimental3DTileContent.fromI3dm = function (
 ) {
   const content = new ModelExperimental3DTileContent(tileset, tile, resource);
 
-  const modelOptions = {
+  const additionalOptions = {
     arrayBuffer: arrayBuffer,
     byteOffset: byteOffset,
     resource: resource,
-    cull: false, // The model is already culled by 3D Tiles
-    releaseGltfJson: true, // Models are unique and will not benefit from caching so save memory
-    opaquePass: Pass.CESIUM_3D_TILE, // Draw opaque portions of the model during the 3D Tiles pass
-    modelMatrix: tile.computedTransform,
-    upAxis: tileset._gltfUpAxis,
-    forwardAxis: Axis.X,
-    incrementallyLoadTextures: false,
-    customShader: tileset.customShader,
-    content: content,
-    colorBlendMode: tileset.colorBlendMode,
-    colorBlendAmount: tileset.colorBlendAmount,
   };
+
+  const modelOptions = makeModelOptions(
+    tileset,
+    tile,
+    content,
+    additionalOptions
+  );
   content._model = ModelExperimental.fromI3dm(modelOptions);
   return content;
 };
@@ -306,10 +294,24 @@ ModelExperimental3DTileContent.fromPnts = function (
 ) {
   const content = new ModelExperimental3DTileContent(tileset, tile, resource);
 
-  const modelOptions = {
+  const additionalOptions = {
     arrayBuffer: arrayBuffer,
     byteOffset: byteOffset,
     resource: resource,
+  };
+
+  const modelOptions = makeModelOptions(
+    tileset,
+    tile,
+    content,
+    additionalOptions
+  );
+  content._model = ModelExperimental.fromPnts(modelOptions);
+  return content;
+};
+
+function makeModelOptions(tileset, tile, content, additionalOptions) {
+  const mainOptions = {
     cull: false, // The model is already culled by 3D Tiles
     releaseGltfJson: true, // Models are unique and will not benefit from caching so save memory
     opaquePass: Pass.CESIUM_3D_TILE, // Draw opaque portions of the model during the 3D Tiles pass
@@ -319,9 +321,17 @@ ModelExperimental3DTileContent.fromPnts = function (
     incrementallyLoadTextures: false,
     customShader: tileset.customShader,
     content: content,
+    show: tileset.show,
     colorBlendMode: tileset.colorBlendMode,
     colorBlendAmount: tileset.colorBlendAmount,
+    lightColor: tileset.lightColor,
+    featureIdLabel: tileset.featureIdLabel,
+    instanceFeatureIdLabel: tileset.instanceFeatureIdLabel,
+    pointCloudShading: tileset.pointCloudShading,
+    backFaceCulling: tileset.backFaceCulling,
+    shadows: tileset.shadows,
+    showCreditsOnScreen: tileset.showCreditsOnScreen,
   };
-  content._model = ModelExperimental.fromPnts(modelOptions);
-  return content;
-};
+
+  return combine(additionalOptions, mainOptions);
+}