+++ Specs/Scene/GlobeSurfaceTileProviderSpec.js
@@ -524,6 +524,43 @@ describe(
         });
       });
 
+      it("culls tiles but does not render fog visuals when renderable is false", function () {
+        expect(scene).toRender([0, 0, 0, 255]);
+        scene.imageryLayers.addImageryProvider(
+          new SingleTileImageryProvider({
+            url: "Data/Images/Red16x16.png",
+          })
+        );
+        const oldFog = scene.fog;
+        scene.fog = new Fog();
+        switchViewMode(
+          SceneMode.SCENE3D,
+          new GeographicProjection(Ellipsoid.WGS84)
+        );
+        scene.camera.lookUp(1.2); // Horizon-view
+
+        return updateUntilDone(scene.globe).then(function () {
+          expect(scene).notToRender([0, 0, 0, 255]);
+
+          scene.fog.enabled = true;
+          scene.fog.renderable = false;
+          scene.fog.density = 0.001;
+          scene.fog.screenSpaceErrorFactor = 0.0;
+
+          let result;
+          expect(scene).toRenderAndCall(function (rgba) {
+            result = rgba;
+            expect(rgba).not.toEqual([0, 0, 0, 255]);
+          });
+
+          scene.fog.renderable = false;
+          expect(scene).notToRender(result);
+          expect(scene).notToRender([0, 0, 0, 255]);
+
+          scene.fog = oldFog;
+        });
+      });
+
       it("culls tiles because of increased SSE", function () {
         expect(scene).toRender([0, 0, 0, 255]);
         scene.imageryLayers.addImageryProvider(
@@ -848,11 +885,11 @@ describe(
         for (const tileID in drawCommandsPerTile) {
           if (drawCommandsPerTile.hasOwnProperty(tileID)) {
             ++tileCount;
-            expect(drawCommandsPerTile[tileID]).toBeGreaterThanOrEqualTo(2);
+            expect(drawCommandsPerTile[tileID]).toBeGreaterThanOrEqual(2);
           }
         }
 
-        expect(tileCount).toBeGreaterThanOrEqualTo(1);
+        expect(tileCount).toBeGreaterThanOrEqual(1);
       });
     });